AnnapurnaAI/annapurna_ai_backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
scikit-learn>=1.3.0
pandas>=2.1.0
joblib>=1.3.0
numpy>=1.24.0